+++ src/components/ManaCalculator/actions/calculateTPS.ts
@@ -1,5 +1,5 @@
-import { EPOCH_DURATION } from "../constants";
+import { EPOCH_DURATION } from '../constants';
 
 export function calculateTPS(mana: number, congestion: number): number {
-    return mana / congestion / EPOCH_DURATION;
-  }
+  return mana / congestion / EPOCH_DURATION;
+}
