+++ src/components/ManaCalculator/enums/index.ts
@@ -1,1 +1,1 @@
-export * from './parameters.enum'
+export * from './parameters.enum';
