+++ src/components/ManaCalculator/types/index.ts
@@ -1,3 +1,2 @@
-export * from './events.type'
-export * from './mana-calculator.type'
-export * from './validator.type'
+export * from './mana-calculator.type';
+export * from './validator.type';
